+++ advent/Days/2017/Day05.cs
@@ -38,7 +38,7 @@ namespace advent.Days._2017
             return NavigateList(input, x=> x + 1);
         }
 
-        public override int Part2(string input)
+        public override long Part2(string input)
         {
             return NavigateList(input, x =>
                 x >= 3
