+++ advent/Days/2017/Day08.cs
@@ -82,7 +82,7 @@ namespace advent.Days._2017
             return Part0(input, (x, y) => x);
         }
 
-        public override int Part2(string input)
+        public override long Part2(string input)
         {
             return Part0(input, (x, y) => y);
         }
