+++ advent/Days/2017/Day13.cs
@@ -47,7 +47,7 @@ namespace advent.Days._2017
             return severity;
         }
 
-        public override int Part2(string input)
+        public override long Part2(string input)
         {
             var firewall = Part0(input);
