+++ advent/Days/2020/Day04.cs
@@ -6,6 +6,6 @@ namespace advent.Days._2020
     public class Day04 : Day
     {
         public override int Part1(string input) => throw new NotImplementedException();
-        public override int Part2(string input) => throw new NotImplementedException();
+        public override long Part2(string input) => throw new NotImplementedException();
     }
 }
