+++ advent/Days/2020/Day05.cs
@@ -5,6 +5,6 @@ namespace advent.Days._2020
     public class Day05 : Day
     {
         public override int Part1(string input) => throw new NotImplementedException();
-        public override int Part2(string input) => throw new NotImplementedException();
+        public override long Part2(string input) => throw new NotImplementedException();
     }
 }
