+++ advent/Days/Day.cs
@@ -7,6 +7,6 @@ namespace advent.Days
         public IConsole Console { get; set; } = new AdventConsole();
 
         public abstract int Part1(string input);
-        public abstract int Part2(string input);
+        public abstract long Part2(string input);
     }
 }
