+++ advent/Days/IDay.cs
@@ -3,6 +3,6 @@
     public interface IDay
     {
         int Part1(string input);
-        int Part2(string input);
+        long Part2(string input);
     }
 }
